api: avoid overwriting mode when falling back
When a player doesn't have data for a match type, we fall back to QM or RM respectively.
However, if we overwrite the original mode, that means that for the next player we check,
we will start looking for the wrong mode rather than attempting to grab the correct mode
first. We can avoid this problem by copying mode into another variable and using that for
our queries.